-- ITERATION 02 : mise en place la BLL et tests mockés--
S008 - Couche BLL : ajouter un client et son adresse

S009 - Couche BLL :  trouver un client par nom

S010 - Couche BLL : trouver les clients par nom

S011 - Couche BLL : Modification complète : nom, prenom, email, téléphone et adresse

S012 - Couche BLL : Modifier uniquement l'adresse d'un client
+ Tests de l'ajout du client (DAL mockée)

S013 - Couche BLL : CRUD du jeu (et des liens vers les genres)
+ Tests de l'ajout d'un jeu (DAL mockée)

S014 - Couche BLL : Consulter la liste des jeux
- le nombre d'exemplaire(s) de chaque jeu doit être indiqué
[- il doit être possible de filtrer par titre (contient) et/ou par genres ]
+ Tests

S015 - Couche BLL : Sortir un exemplaire (<=>Créer une location)
+ Tests

S016 - Couche BLL : Retourner une liste d'exemplaires
( et créer la facture )
+ Tests

S017 - Couche BLL : Marquer la facture comme payée
+ Tests

S018 - Couche BLL : Modifier l'état d'un exemplaire
+ Tests

+++ src/main/java/dubbelf/fr/ludotechtp/bo/Client.java
@@ -12,7 +12,7 @@ public class Client {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long noClient;
+    private Integer noClient;
 
     @Column(nullable = false, length = 50)
     @NonNull
